refactor: controller 분리 및 swagger 주석 추가
