+++ src/main/java/cart/ui/controller/PageController.java
@@ -7,11 +7,11 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class PageApiController {
+public class PageController {
     private final ProductService productService;
     private final MemberDao memberDao;
 
-    public PageApiController(ProductService productService, MemberDao memberDao) {
+    public PageController(ProductService productService, MemberDao memberDao) {
         this.productService = productService;
         this.memberDao = memberDao;
     }
